fight warnings c++17 error throw(). Build GEOM finished. Fix other warnings

+++ src/Container/TestSalome_file.cxx
@@ -73,7 +73,7 @@ int main (int argc, char * argv[])
   std::cerr << "+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++" << std::endl;
   std::cerr << "Test of getFilesInfos()" << std::endl;
   all_infos = file.getFilesInfos();
-  for (int i = 0; i < all_infos->length(); i++)
+  for (int i = 0; i < (int)all_infos->length(); i++)
   {
     print_infos(&((*all_infos)[i]));
   }
